BST update
