+++ Data/BST.java
@@ -2,39 +2,19 @@ package Data;
 import java.util.Scanner;
 
 public class BST<T> {
-
-    static class Node<T> {
-        public T key;
-        public Node<Integer> right;
-        public Node<Integer> left;
-
-        // Const.
-        public Node(T data) {
-            super();
-            this.key = data;
-            this.right = null;
-            this.left = null;
-        }
-
-        @Override
-        public String toString(){
-            return String.valueOf(this.key);
-        }
-    }
-
-    private Node<Integer> root = null;
+    private NodeT<T> root = null;
 
     public BST() {
         this.root = null;
     }
 
     public void insert (int num){
-        this.insert(num, (Node<Integer>) this.root);
+        this.insert(num,this.root);
     }
 
-    public Node<Integer> insert(int num,Node<Integer> ptr) {
+    public NodeT<Integer> insert(int num,NodeT<T> ptr) {
         if (ptr==null) {
-            ptr = new Node<Integer>(num);
+            ptr = new NodeT<Integer>(num);
         } else {
             if (num < ptr.key) {
                 ptr.left = insert(num, ptr.left);
@@ -65,46 +45,9 @@ public class BST<T> {
         return this.root == null;
     }
 
-    private String pathBreadth(Queue<BST.Node<T>> level) {
-        String result = "";
-        boolean flag = true;
-        if (!this.isEmpty()) {
-            if (level.isEmpty()) {
-                level.enqueue(new Data.Node<Node<T>>((Node<T>) this.root));
-                return result += pathBreadth(level);
-            } else {
-                Node<T> cur = level.dequeue();
-                if (cur == null) {
-                    result += "-";
-                    flag = false;
-                } else {
-                    result += cur.toString();
-                }
-                if (flag) {
-                    if (cur.left != null) {
-                        level.enqueue(new Data.Node<Node<T>>((Node<T>) cur.left));
-                    } else {
-                        level.enqueue(new Data.Node<Node<T>>(new Node<T>(null)));
-                    }
-                    if (cur.right != null) {
-                        level.enqueue(new Data.Node<Node<T>>((Node<T>) cur.right));
-                    } else {
-                        level.enqueue(new Data.Node<Node<T>>(new Node<T>(null)));
-                    }
-                }
-                if (!level.isEmpty()) {
-                    return result += " " + pathBreadth(level);
-                } else {
-                    return result;
-                }
-            }
-        }
-        return result;
-    }
-
-    public String levelOrder() {
-        Queue<BST.Node<T>> level = new Queue<BST.Node<T>>();
-        return this.pathBreadth(level);
+    public String levelOrder(){
+        Queue<NodeT<T>> level = new Queue<NodeT<T>>();
+        level.enqueue(new Node<NodeT<T>>(this.root));
     }
 
     public static void main(String[] args) {
@@ -114,6 +57,8 @@ public class BST<T> {
         Scanner input = new Scanner(System.in);
         String data = input.nextLine();
         myTree.inputLineTOInsert(data);
-        System.out.println(myTree.levelOrder());
+        System.out.println(myTree.root.key);
+        System.out.println(myTree.root.right.key);
+        System.out.println(myTree.root.left.key);
     }
 }
