+++ Data/NodeT.java
@@ -1,0 +1,33 @@
+package Data;
+public class NodeT<T> {
+    // Att.
+    public T key;
+    public NodeT<T> right;
+    public NodeT<T> left;
+
+    // Const.
+    public NodeT(T data) {
+        this.key = data;
+        this.right = null;
+        this.left = null;
+    }
+
+    // Meth.
+
+    public T getData() {
+        return this.key;
+    }
+
+    public void updateKey(T data) {
+        this.key = data;
+    }
+
+    @Override
+    public String toString(){
+        if (this.key==null){
+            return "-";
+        } else {
+            return String.valueOf(this.key);
+        }
+    }
+}
